refactor: restructure collector into submodules
Move InMemoryCollector to collector/memory.rs, rename file_collector
to collector/file.rs, keep shared tests in collector/mod.rs.
